+++ anilist/find.go
@@ -1,12 +1,27 @@
 package anilist
 
 import (
+	"fmt"
 	levenshtein "github.com/ka-weihe/fast-levenshtein"
 	"github.com/metafates/mangal/log"
+	"github.com/metafates/mangal/util"
 	"github.com/samber/lo"
+	"strings"
+)
+
+var (
+	retries uint8
+	limit   uint8 = 3
 )
 
 func FindClosest(name string) (*Manga, error) {
+	if retries >= limit {
+		retries = 0
+		err := fmt.Errorf("no results found on Anilist for manga %s", name)
+		log.Error(err)
+		return nil, err
+	}
+
 	if cached, ok := cache[name]; ok {
 		log.Info("Found cached manga: " + cached.Name())
 		return cached, nil
@@ -20,8 +35,12 @@ func FindClosest(name string) (*Manga, error) {
 	}
 
 	if len(urls) == 0 {
-		log.Warn("No results found on Anilist for manga" + name)
-		return nil, nil
+		// try again with a different name
+		retries++
+		words := strings.Split(name, " ")
+		// one word less
+		alternateName := strings.Join(words[:util.Max(len(words)-1, 0)], " ")
+		return FindClosest(alternateName)
 	}
 
 	// find the closest match
@@ -31,5 +50,6 @@ func FindClosest(name string) (*Manga, error) {
 
 	log.Info("Found closest match: " + closest.Name())
 	cache[name] = closest
+	retries = 0
 	return closest, nil
 }