+++ constant/build.go
@@ -1,0 +1,8 @@
+package constant
+
+const Version = "3.8.0"
+
+var (
+	Built     string
+	GitCommit string
+)